refactor : 반복되는 querySelector 함수화하여 리팩토링
