+++ src/js/view/inputView.js
@@ -1,21 +1,22 @@
 import { selectors } from "../keys.js";
+import { $ } from "../utils.js";
 
 export const initializeInputView = function () {
-  document.querySelector(selectors.carNamesInput).value = "";
-  document.querySelector(selectors.countInput).value = "";
-  document.querySelector(selectors.countContainer).style.display = "none";
+  $(selectors.carNamesInput).value = "";
+  $(selectors.countInput).value = "";
+  $(selectors.countContainer).style.display = "none";
 };
 
 export const displayCountView = function () {
-  document.querySelector(selectors.countContainer).style.display = "block";
+  $(selectors.countContainer).style.display = "block";
 };
 
 export const toggleCarNameInputDisable = function () {
-  document.querySelector(selectors.carNamesInput).toggleAttribute("disabled");
-  document.querySelector(selectors.carNamesSubmit).toggleAttribute("disabled");
+  $(selectors.carNamesInput).toggleAttribute("disabled");
+  $(selectors.carNamesSubmit).toggleAttribute("disabled");
 };
 
 export const toggleCountInputDisable = function () {
-  document.querySelector(selectors.countInput).toggleAttribute("disabled");
-  document.querySelector(selectors.countSubmit).toggleAttribute("disabled");
+  $(selectors.countInput).toggleAttribute("disabled");
+  $(selectors.countSubmit).toggleAttribute("disabled");
 };
