+++ src/js/view/winnerView.js
@@ -1,13 +1,14 @@
 import { selectors, globalTexts } from "../keys.js";
+import { $ } from "../utils.js";
 
 export const displayWinnerView = function (winners) {
-  const $winnerElement = document.querySelector(selectors.winnerTextArea);
+  const $winnerElement = $(selectors.winnerTextArea);
   $winnerElement.innerText = globalTexts.makeWinnerText(winners);
-  document.querySelector(selectors.winnerContainer).style.display = "flex";
+  $(selectors.winnerContainer).style.display = "flex";
 };
 
 export const initializeWinnerView = function () {
-  const $winnerElement = document.querySelector(selectors.winnerTextArea);
-  document.querySelector(selectors.winnerContainer).style.display = "none";
+  const $winnerElement = $(selectors.winnerTextArea);
+  $(selectors.winnerContainer).style.display = "none";
   $winnerElement.innerText = globalTexts.winnerText;
 };
